broke react-tinacms into multiple files
