+++ packages/react/react-tinacms/src/index.ts
@@ -1,2 +1,7 @@
-export * from './cms-react'
 export * from './Form'
+export * from './use-cms'
+export * from './use-form'
+export * from './use-plugin'
+export * from './use-subscribable'
+export * from './watch-form-values'
+export * from './with-plugin'
