+++ packages/react/react-tinacms/src/use-form.ts
@@ -1,0 +1,29 @@
+import { FormOptions, Form } from '@tinacms/core'
+import * as React from 'react'
+import { useCMS } from './use-cms'
+
+export function useCMSForm(options: FormOptions<any>) {
+  let cms = useCMS()
+  let [form, setForm] = React.useState<Form | null>(null)
+  let [_, setValues] = React.useState(options.initialValues)
+
+  React.useEffect(
+    function createForm() {
+      let form = cms.forms.createForm(options)
+      setForm(form)
+      form.subscribe(
+        form => {
+          setValues(form.values)
+        },
+        { values: true }
+      )
+
+      return () => {
+        form && cms.forms.removeForm(form.name)
+      }
+    },
+    [options.name]
+  )
+
+  return [form ? form.values : options.initialValues, form]
+}
